% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_call.R
\name{batch_call}
\alias{batch_call}
\title{Breaks longer API calls into batches so that it can manage.}
\usage{
batch_call()
}
\arguments{
\item{vector_of_dates}{A list of raster in velox format. Shoudl be in YYYY-MM-DD format}
}
\value{
An spdf with the same format as the velox raster data. We can re-use the dates list again without passing it through the function as returned output.
}
\description{
Breaks longer API calls into batches so that it can manage.
}
\note{
This might not be necessary if we can get this set up on a server.
}
\examples{
arc2_api_download(c("2019-01-01", "2019-01-02"))
}
