% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_directory.R
\name{clean_data_directory}
\alias{clean_data_directory}
\title{clean raw data directory of dated files.}
\usage{
clean_data_directory(dir, pattern = "weatherRasterList")
}
\arguments{
\item{dir}{directory where raw data lives}

\item{listToAccess}{the list of files to access from get_new_additions}
}
\value{
nothing. It saves a single data file and deletes the other files.
}
\description{
clean raw data directory of dated files.
}
