% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_data.R
\name{combine_all_data}
\alias{combine_all_data}
\title{Combines all accessed data into a single list of rasters}
\usage{
combine_all_data(directory = NULL)
}
\arguments{
\item{directory}{Where to find the lists of rasters to combine into a single list}
}
\value{
A list of rasters to put into the convert_to_velox sequence to get it ready for extraction
}
\description{
Combines all accessed data into a single list of rasters
}
