% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bin_to_raster.R
\name{convert_bin_to_raster}
\alias{convert_bin_to_raster}
\title{convert ftp data to matrix and then raster}
\usage{
convert_bin_to_raster(binaryInput, fileUrl)
}
\arguments{
\item{binaryInput}{Binary input from FTP site}

\item{fileUrl}{web url with location of file}
}
\value{
matrix of values to be converted to raster
}
\description{
convert ftp data to matrix and then raster
}
