% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spdf.R
\name{convert_spdf}
\alias{convert_spdf}
\title{Convert data.frame of GPS points in decimal degree format to an SpatialPointsDataFrame}
\usage{
convert_spdf(df, lat = NULL, lon = NULL, defaultCRS = NULL)
}
\arguments{
\item{df}{A data frame of GPS points in decimal degree format}

\item{lat}{The variable that has the Latitude data}

\item{lon}{The variable that has the Longitude data}

\item{defaultCRS}{The default CRS to use. If none is provided then it uses the default latlon CRS.}
}
\value{
A spdf with a default CRS. This default CRS will then need to be converted.
}
\description{
Convert data.frame of GPS points in decimal degree format to an SpatialPointsDataFrame
}
\examples{
add(1, 1)
add(10, 1)
}
