% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tibble_to_raster.R
\name{convert_tibble_to_raster}
\alias{convert_tibble_to_raster}
\title{Convert tibble data from api call to raster}
\usage{
convert_tibble_to_raster(raw_data_list)
}
\arguments{
\item{list_of_tibble}{A list of tibbles from the API call}
}
\value{
List of raster data. This goes into the convert_to_velox function so needs to fit nicely with that function's expectations
}
\description{
Convert tibble data from api call to raster
}
\examples{
add(1, 1)
add(10, 1)
}
