% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_velox.R
\name{convert_to_velox}
\alias{convert_to_velox}
\title{Make sure the velox raster and the GPS spdf have the same CRS}
\usage{
convert_to_velox(rList)
}
\arguments{
\item{rList}{A list of raster in raster format.}
}
\value{
List of two elements - first is the raster data in velox format. Second are the labels needed to identify the data later.
}
\description{
Make sure the velox raster and the GPS spdf have the same CRS
}
\examples{
add(1, 1)
add(10, 1)
}
