% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_vector.R
\name{create_date_vector}
\alias{create_date_vector}
\title{Create date vector for given start and end date}
\usage{
create_date_vector(start_date, end_date)
}
\arguments{
\item{start_date}{start date, formatted YYYY-MM-DD}

\item{end_date}{end date, formatted YYYY-MM-DD}
}
\value{
A data.frame with the extracted weather values and the date for each GPS point in the file.
}
\description{
Create date vector for given start and end date
}
\examples{
add(1, 1)
add(10, 1)
}
