% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_raw_data.R
\name{create_master_raw_data}
\alias{create_master_raw_data}
\title{consolidate raw data into single file to simplify storage and loading}
\usage{
create_master_raw_data(dir, pattern = "weatherRasterList")
}
\arguments{
\item{dir}{directory where raw data lives. Dir is the raw_data folder from
inhereted from the get_raw_data}

\item{pattern}{the pattern that identifies the raw data files}
}
\value{
nothing. It saves a single data file and deletes the other files.
}
\description{
consolidate raw data into single file to simplify storage and loading
}
