% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_df_creator.R
\name{date_df_creator}
\alias{date_df_creator}
\title{Converts data vector from raw raster file name into useable date information for identification}
\usage{
date_df_creator(dateVector)
}
\arguments{
\item{dateVector}{The date vector from the Arc2 raw weather data}
}
\value{
data.frame of date vector, year, month, and day in numeric format.
}
\description{
Converts data vector from raw raster file name into useable date information for identification
}
\examples{
add(1, 1)
add(10, 1)
}
