% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_velox_gps.R
\name{extract_velox_gps}
\alias{extract_velox_gps}
\title{Extracts values for GPS df from velox raster}
\usage{
extract_velox_gps(veloxRaster, spdf)
}
\arguments{
\item{veloxRaster}{A list of two elements (from convert_to_velox function). First is list of velox raster, second is df of date information.}

\item{spdf}{A SpatialPointsDataFrame of the GPS for which we'll be extracting data from the velox raster.}
}
\value{
A data.frame that is long by date. It has the observation for each date and GPS point in the raster data.
}
\description{
Extracts values for GPS df from velox raster
}
