% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weather_data.R
\name{extract_weather_data}
\alias{extract_weather_data}
\title{Combines the full functionality to take the raw inputs and produce data.frame of extracted weather data and dates.}
\usage{
extract_weather_data(start_date, end_date, gpsFile, latCol, lonCol)
}
\arguments{
\item{start_date}{start date, formatted YYYY-MM-DD}

\item{end_date}{end date, formatted YYYY-MM-DD}

\item{gpsFile}{GPS file, probably from data warehouse, of GPS points for which we want to extract weather values}

\item{latCol}{The Latitude column in the gpsFile}

\item{lonCol}{The Longitude column in the gpsFile}

\item{rawRasterData}{list of raw raster data from Arc2 weather site}
}
\value{
A data.frame with the extracted weather values and the date for each GPS point in the file.
}
\description{
Combines the full functionality to take the raw inputs and produce data.frame of extracted weather data and dates.
}
\examples{
weatherValues <- extract_weather_data(dates, gpsData, "Latitude", "Longitude")
}
