% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_list.R
\name{get_full_list}
\alias{get_full_list}
\title{pulls full list of available datasets from Arc2 weather FTP}
\usage{
get_full_list()
}
\arguments{
\item{url}{web url of FTP}
}
\value{
A vector of the file names available on the FTP
}
\description{
pulls full list of available datasets from Arc2 weather FTP
}
