% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_additions.R
\name{get_new_additions}
\alias{get_new_additions}
\title{creates list of new files to be downloaded, difference between available and current}
\usage{
get_new_additions(existingList, fullList, directory)
}
\arguments{
\item{existingList}{files we've already downloaded}

\item{fullList}{full list of files available on FTP}

\item{directory}{where to find the current files and where to save data}
}
\value{
A vector of the file names to download
}
\description{
creates list of new files to be downloaded, difference between available and current
}
