% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{pulls the yet to be accessed new data from the FTP}
\usage{
get_raw_data(listToAccess, dir, url = NULL)
}
\arguments{
\item{listToAccess}{the list of files to access from get_new_additions}

\item{url}{web url of FTP with the file names}
}
\value{
a list of rasters accessed from the FTP. Saves this file to the directory
}
\description{
pulls the yet to be accessed new data from the FTP
}
