% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arc_binary.R
\name{read_arc_binary}
\alias{read_arc_binary}
\title{reads in binary file from arc2 weather ftp}
\usage{
read_arc_binary(fileUrl)
}
\arguments{
\item{fileUrl}{web location of ftp}
}
\value{
data.frame with the weather data
}
\description{
reads in binary file from arc2 weather ftp
}
\examples{
add(1, 1)
add(10, 1)
}
