% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_directory.R
\name{set_data_directory}
\alias{set_data_directory}
\title{This function sets the directory to the raw data location to simplify
extending data directories in the read and write process}
\usage{
set_data_directory(dir)
}
\arguments{
\item{fileDirectory}{the file location locally where the data names are saved}
}
\value{
A vector of the file names available already
}
\description{
This function sets the directory to the raw data location to simplify
extending data directories in the read and write process
}
