% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_list_directory.R
\name{set_list_directory}
\alias{set_list_directory}
\title{this function checks the directory provided as the location of the ARC2
weather data. It then tries to guess at a reasonable modification but if it
doesn't work it ask for a new directory.}
\usage{
set_list_directory(dir)
}
\arguments{
\item{fileDirectory}{the file location locally where the data names are saved}
}
\value{
A vector of the file names available already
}
\description{
this function checks the directory provided as the location of the ARC2
weather data. It then tries to guess at a reasonable modification but if it
doesn't work it ask for a new directory.
}
