% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_month.R
\name{subset_month}
\alias{subset_month}
\title{Subset the raster list by month}
\usage{
subset_month(rList, month)
}
\arguments{
\item{rList}{A list of raster in raster format.}

\item{month}{A numeric month between 1 and 12 to indicate which month to keep. This can be a vector or range as well.}
}
\value{
List of raster subest by month
}
\description{
Subset the raster list by month
}
\examples{
add(1, 1)
add(10, 1)
}
