% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_year.R
\name{subset_year}
\alias{subset_year}
\title{Subset the raster list by year}
\usage{
subset_year(rList, year)
}
\arguments{
\item{rList}{A list of raster in raster format.}

\item{year}{A numeric year between 1985 and present to indicate which years to keep. This can be a vector or range as well.}
}
\value{
A list of rasters subset by year
}
\description{
Subset the raster list by year
}
\examples{
add(1, 1)
add(10, 1)
}
