% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_year_month.R
\name{subset_year_month}
\alias{subset_year_month}
\title{Subset the raster list by month and year combining the subset month and year functions}
\usage{
subset_year_month(rList, year = NULL, month = NULL)
}
\arguments{
\item{rList}{A list of raster in raster format.}

\item{year}{A numeric year between 1985 and present. A range is also acceptable.}

\item{month}{A numeric month between 1 and 12 to indicate which month to keep. This can be a vector or range as well.}
}
\value{
A list of rasters for the requested months and years
}
\description{
Subset the raster list by month and year combining the subset month and year functions
}
\examples{
add(1, 1)
add(10, 1)
}
