% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_arc_weather_data.R
\name{tidy_arc_weather_data}
\alias{tidy_arc_weather_data}
\title{Applies some general cleaning to the extracted data before exporting}
\usage{
tidy_arc_weather_data(df)
}
\arguments{
\item{df}{The data.frame coming from the full extract process.}
}
\value{
A data.frame that has some variables cleaned and labels clarified.
}
\description{
Applies some general cleaning to the extracted data before exporting
}
\examples{
add(1, 1)
add(10, 1)
}
