% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_arc_weather_data.R
\name{update_arc_weather_data}
\alias{update_arc_weather_data}
\title{combines the data download functions into a single function}
\usage{
update_arc_weather_data(dir)
}
\arguments{
\item{dir}{the directory where to find the files already downloaded}

\item{url}{web url of FTP}
}
\value{
list of rasters downloaded from the FTP
}
\description{
combines the data download functions into a single function
}
