% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_current_list.R
\name{update_current_list}
\alias{update_current_list}
\title{takes the list of file names and creates a single master list that will be referenced and updated going forward.}
\usage{
update_current_list(directory = NULL)
}
\arguments{
\item{directory}{the location where those files live}
}
\value{
an rds file that has all the data that has been accessed up to this point
}
\description{
takes the list of file names and creates a single master list that will be referenced and updated going forward.
}
