% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_list.R
\name{get_current_list}
\alias{get_current_list}
\title{gets names of current files already downloaded and worked with}
\usage{
get_current_list(fileDirectory = NULL)
}
\arguments{
\item{fileDirectory}{the file location locally where the data names are saved}
}
\value{
A vector of the file names available already
}
\description{
gets names of current files already downloaded and worked with
}
