% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arc_binary.R
\name{read_arc_binary}
\alias{read_arc_binary}
\title{reads in binary file from arc2 weather ftp}
\usage{
read_arc_binary(fileUrl)
}
\arguments{
\item{fileUrl}{thee full web location of ftp. This funciton lies in the
get_raw_data structure deliberately so that we are pulling the data with
the full url.}
}
\value{
weather data in binary format. This function is nested in
  convert_bin_to_ratster and get_raw_data to go from binary to df.
}
\description{
reads in binary file from arc2 weather ftp
}
\examples{
add(1, 1)
add(10, 1)
}
